+++ bracketingTour.cpp
@@ -1,0 +1,10 @@
+#include "bracketingTour.h"
+
+int BracketingTour::DoTour(){
+    //* Insert code here to create the list for bracketing rankings
+    // check for buy.
+    
+}
+int BracketingTour::InitializeTour(){
+    //* Insert code here to initialize bracketing rankings
+}
