+++ bracketingTour.h
@@ -1,0 +1,16 @@
+#ifndef BRACKETINGTOUR_H
+#define BRACKETINGTOUR_H
+
+#include "tournament.h"
+class BracketingTour: public Tournament
+{
+    private:
+
+    public:
+    int     InitializeTour  ();
+    int     DoTour          ();
+};
+
+
+
+#endif
